Get data by unit Id's from BDL API

Retrieve data for given units from BDL with specified format.

get_data_by_unit(
  unitId,
  varId,
  year = NULL,
  type = c("code", "label"),
  aggregateId = NULL,
  lang = c("pl", "en"),
  ...
)

Arguments

unitId
A single 12 character NUTS id code or vector of multiple unit id codes. If multiple unit codes are used, some columns are not available. Use search_units or get_units to find unit id code.

varId
A vector of variable Id's.Use search_variables or get_variables to find variable id code.

year
A vector of years. If NULL (default) returns data for all available years.

type
A type of variables returned, "code" (default), "label"

aggregateId
An aggregate id. Use get_aggregates for more info.

lang
A language of returned data, "pl" (default), "en"

...
Other arguments passed on to GET. For example a proxy parameters, see details.

Value

A dataset as a tibble.

Details

Data to retrieve from The BDL Web Services can be filtered with arguments. To get JSON data from specified directory with custom filters use get_request directly.

To use a proxy to connect, a use_proxy can be passed to GET. For example get_request(id, filters, config = httr::use_proxy(url, port, username, password)).

Examples

 # \donttest{
   get_data_by_unit(unitId = "023200000000", varId =  "3643")

#> # A tibble: 20 x 6
#>    id    measureUnitId lastUpdate              year    val attrId
#>    <chr>         <int> <chr>                   <chr> <dbl>  <int>
#>  1 3643             26 2019-05-22T16:15:20.397 1999      1      1
#>  2 3643             26 2019-05-22T16:15:20.397 2000      2      1
#>  3 3643             26 2019-05-22T16:15:20.397 2001      4      1
#>  4 3643             26 2019-05-22T16:15:20.397 2002      3      1
#>  5 3643             26 2019-05-22T16:15:20.397 2003      3      1
#>  6 3643             26 2019-05-22T16:15:20.397 2004      2      1
#>  7 3643             26 2019-05-22T16:15:20.397 2005      3      1
#>  8 3643             26 2019-05-22T16:15:20.397 2006      4      1
#>  9 3643             26 2019-05-22T16:15:20.397 2007      5      1
#> 10 3643             26 2019-05-22T16:15:20.397 2008      5      1
#> 11 3643             26 2019-05-22T16:15:20.397 2009      5      1
#> 12 3643             26 2019-05-22T16:15:20.397 2010      6      1
#> 13 3643             26 2019-05-22T16:15:20.397 2011      7      1
#> 14 3643             26 2019-05-22T16:15:20.397 2012      4      1
#> 15 3643             26 2019-05-22T16:15:20.397 2013      3      1
#> 16 3643             26 2019-05-22T16:15:20.397 2014      4      1
#> 17 3643             26 2019-05-22T16:15:20.397 2015      3      1
#> 18 3643             26 2019-05-22T16:15:20.397 2016      2      1
#> 19 3643             26 2019-05-22T16:15:20.397 2017      1      1
#> 20 3643             26 2019-05-22T16:15:20.397 2018      6      1

   get_data_by_unit(unitId = "023200000000", varId =  c("3643", "2137", "148190"),
                          type = "label")

#> # A tibble: 50 x 9
#>    id    measureUnitId lastUpdate year    val attrId variableName measureName
#>    <chr>         <int> <chr>      <chr> <dbl>  <int> <chr>        <chr>      
#>  1 2137             18 2019-09-3~ 1998  10402      1 Gospodarka ~ m3         
#>  2 2137             18 2019-09-3~ 1999  95582      1 Gospodarka ~ m3         
#>  3 2137             18 2019-09-3~ 2000  12977      1 Gospodarka ~ m3         
#>  4 2137             18 2019-09-3~ 2001   1486      1 Gospodarka ~ m3         
#>  5 2137             18 2019-09-3~ 2002   5863      1 Gospodarka ~ m3         
#>  6 2137             18 2019-09-3~ 2003   8655      1 Gospodarka ~ m3         
#>  7 2137             18 2019-09-3~ 2004   7686      1 Gospodarka ~ m3         
#>  8 2137             18 2019-09-3~ 2005  31930      1 Gospodarka ~ m3         
#>  9 2137             18 2019-09-3~ 2006  19749      1 Gospodarka ~ m3         
#> 10 2137             18 2019-09-3~ 2007   4933      1 Gospodarka ~ m3         
#> # ... with 40 more rows, and 1 more variable: attributeDescription <chr>

   # Multi variable download
   get_data_by_unit(unitId = c("023200000000", "020800000000"),
                          varId =  c("3643", "2137", "148190"))

#> Joining, by = c("id", "year")

#> # A tibble: 50 x 4
#>    id    year  val_023200000000 val_020800000000
#>    <chr> <chr>            <dbl>            <dbl>
#>  1 2137  1998             10402            10929
#>  2 2137  1999             95582            10125
#>  3 2137  2000             12977             2190
#>  4 2137  2001              1486             2340
#>  5 2137  2002              5863             4398
#>  6 2137  2003              8655             1032
#>  7 2137  2004              7686              598
#>  8 2137  2005             31930             2040
#>  9 2137  2006             19749             6264
#> 10 2137  2007              4933             1854
#> # ... with 40 more rows

# }