Get data by variable Id from BDL API

Retrieve data for a given variable for multiple units from BDL with specified format.

get_data_by_variable(varId, unitParentId = NULL, unitLevel = NULL,
  year = NULL, aggregateId = NULL, lang = c("pl", "en"), ...)

Arguments

varId
A single variable Id or vector of multiple variable id's. If multiple id's are used, some columns are not available. Use search_variables or get_variables to find variable id code.

unitParentId
A 12 character NUTS id code of parent unit. Use search_units or get_units to find unit id code. If NULL (default) and unitLevel not set up, returns all available units for variable.

unitLevel
A number from 0 to 6, filters the returned unit by its level. If NULL (default) no level filters apply. Use get_levels to find more info.

year
A vector of years. If NULL (default) returns data for all available years.

aggregateId
An aggregate id. Use get_aggregates for more info.

lang
A language of returned data, "pl" (default), "en"

...
Other arguments passed on to GET. For example a proxy parameters, see details.

Value

A dataset as a tibble.

Details

Data to retrieve from The BDL Web Services can be filtered with arguments. To get JSON data from specified directory with custom filters use get_request directly.

To use a proxy to connect, a use_proxy can be passed to GET. For example get_request(id, filters, config = httr::use_proxy(url, port, username, password)).

Examples

 

   get_data_by_variable(varId = "3643", unitParentId = "030200000000")

#> # A tibble: 40 x 5
#>    id           name         year    val attrId
#>    <chr>        <chr>        <chr> <dbl>  <int>
#>  1 030200000000 DOLNOŚLĄSKIE 1999      9      1
#>  2 030200000000 DOLNOŚLĄSKIE 2000     10      1
#>  3 030200000000 DOLNOŚLĄSKIE 2001      6      1
#>  4 030200000000 DOLNOŚLĄSKIE 2002      6      1
#>  5 030200000000 DOLNOŚLĄSKIE 2003      5      1
#>  6 030200000000 DOLNOŚLĄSKIE 2004      4      1
#>  7 030200000000 DOLNOŚLĄSKIE 2005      6      1
#>  8 030200000000 DOLNOŚLĄSKIE 2006     10      1
#>  9 030200000000 DOLNOŚLĄSKIE 2007     10      1
#> 10 030200000000 DOLNOŚLĄSKIE 2008     10      1
#> # ... with 30 more rows

   get_data_by_variable("420", year = "2000", unitLevel = 6)

#> # A tibble: 290 x 5
#>    id           name                       year    val attrId
#>    <chr>        <chr>                      <chr> <dbl>  <int>
#>  1 011212006113 Skawina                    2000    117      1
#>  2 011212006114 Skawina - miasto           2000    117      1
#>  3 011212161011 Kraków                     2000  54471      1
#>  4 011212210073 Krynica-Zdrój              2000   5137      1
#>  5 011212210074 Krynica-Zdrój - miasto     2000   5137      1
#>  6 011212262011 Nowy Sącz                  2000   5853      1
#>  7 011212312042 Klucze                     2000    826      1
#>  8 011212404023 Dąbrowa Tarnowska          2000   1812      1
#>  9 011212404024 Dąbrowa Tarnowska - miasto 2000   1812      1
#> 10 011212416103 Tuchów                     2000    964      1
#> # ... with 280 more rows

   # Multi variable download
   get_data_by_variable(varId =c("3643","420"), unitParentId = "010000000000")

#> Joining, by = c("id", "name", "year")

#> # A tibble: 100 x 5
#>    id           name                   year  val_3643 val_420
#>    <chr>        <chr>                  <chr>    <dbl>   <dbl>
#>  1 010000000000 MAKROREGION POŁUDNIOWY 1999        12  225459
#>  2 010000000000 MAKROREGION POŁUDNIOWY 2000         9  129602
#>  3 010000000000 MAKROREGION POŁUDNIOWY 2001        10  140821
#>  4 010000000000 MAKROREGION POŁUDNIOWY 2002         8   78668
#>  5 010000000000 MAKROREGION POŁUDNIOWY 2003         7   66702
#>  6 010000000000 MAKROREGION POŁUDNIOWY 2004         5   44721
#>  7 010000000000 MAKROREGION POŁUDNIOWY 2005         4   41309
#>  8 010000000000 MAKROREGION POŁUDNIOWY 2006        12   79826
#>  9 010000000000 MAKROREGION POŁUDNIOWY 2007        10   11746
#> 10 010000000000 MAKROREGION POŁUDNIOWY 2008         8   37162
#> # ... with 90 more rows