Get panel data by unit and variable Id's from BDL API

Retrieve data for given units from BDL with specified format.

get_panel_data(unitId, varId, year = NULL, ggplot = FALSE, ...)

Arguments

unitId
A single 12 character NUTS id code or vector of multiple unit id codes. If multiple unit codes are used, some columns are not available. Use search_units or get_units to find unit id code.

varId
A single Id or vector of variable Id's.Use search_variables or get_variables to find variable id code.

year
A vector of years. If NULL (default) returns data for all available years.

ggplot
Output in a long format suitable for ggplot2. Allows to plot results directly with ggplot function.

...
Other arguments passed on to GET. For example a proxy parameters, see details.

Value

A dataset as a tibble.

Details

Data to retrieve from The BDL Web Services can be filtered with arguments. To get JSON data from specified directory with custom filters use get_request directly.

To use a proxy to connect, a use_proxy can be passed to GET. For example get_request(id, filters, config = httr::use_proxy(url, port, username, password)).

Examples

 # \donttest{
   get_panel_data(unitId = "030210101000", varId =  "60270")

#> # A tibble: 15 x 3
#>    unit         year  `60270`
#>    <fct>        <chr>   <dbl>
#>  1 030210101000 2004     25.3
#>  2 030210101000 2005     22.4
#>  3 030210101000 2006     17.7
#>  4 030210101000 2007     13.4
#>  5 030210101000 2008     11.3
#>  6 030210101000 2009     15.8
#>  7 030210101000 2010     14.9
#>  8 030210101000 2011     12.2
#>  9 030210101000 2012     13.9
#> 10 030210101000 2013     12.6
#> 11 030210101000 2014      8.9
#> 12 030210101000 2015      7.6
#> 13 030210101000 2016      5.7
#> 14 030210101000 2017      4.2
#> 15 030210101000 2018      3.8

   get_panel_data(unitId = "030210101000", varId =  c("60270", "461668"))

#> # A tibble: 16 x 4
#>    unit         year  `60270` `461668`
#>    <fct>        <chr>   <dbl>    <dbl>
#>  1 030210101000 2004     25.3     NA  
#>  2 030210101000 2005     22.4     NA  
#>  3 030210101000 2006     17.7     NA  
#>  4 030210101000 2007     13.4     NA  
#>  5 030210101000 2008     11.3     NA  
#>  6 030210101000 2009     15.8     NA  
#>  7 030210101000 2010     14.9     NA  
#>  8 030210101000 2011     12.2     NA  
#>  9 030210101000 2012     13.9     NA  
#> 10 030210101000 2013     12.6      2.6
#> 11 030210101000 2014      8.9      2.1
#> 12 030210101000 2015      7.6      1.9
#> 13 030210101000 2016      5.7      1.5
#> 14 030210101000 2017      4.2      1.3
#> 15 030210101000 2018      3.8      1.1
#> 16 030210101000 2019     NA        0.9

   get_panel_data(unitId = c("030210101000", "030210105000", "030210106000"),
                          varId =  c("60270", "461668"), year = c(2013:2016))

#> # A tibble: 12 x 4
#>    unit         year  `60270` `461668`
#>    <fct>        <chr>   <dbl>    <dbl>
#>  1 030210101000 2013     12.6      2.6
#>  2 030210101000 2014      8.9      2.1
#>  3 030210101000 2015      7.6      1.9
#>  4 030210101000 2016      5.7      1.5
#>  5 030210105000 2013     24.3      3.6
#>  6 030210105000 2014     18.6      3.3
#>  7 030210105000 2015     17.3      3.7
#>  8 030210105000 2016     15.8      2.9
#>  9 030210106000 2013     19.8      2.3
#> 10 030210106000 2014     15.7      1.9
#> 11 030210106000 2015     12.9      1.6
#> 12 030210106000 2016     11.4      1.3

   get_panel_data(unitId = c("030210101000", "030210105000", "030210106000"),
                          varId =  c("60270", "461668"), ggplot = TRUE)

#> # A tibble: 96 x 4
#>    unit         year  variables values
#>    <fct>        <chr> <chr>      <dbl>
#>  1 030210101000 2004  60270       25.3
#>  2 030210101000 2004  461668      NA  
#>  3 030210101000 2005  60270       22.4
#>  4 030210101000 2005  461668      NA  
#>  5 030210101000 2006  60270       17.7
#>  6 030210101000 2006  461668      NA  
#>  7 030210101000 2007  60270       13.4
#>  8 030210101000 2007  461668      NA  
#>  9 030210101000 2008  60270       11.3
#> 10 030210101000 2008  461668      NA  
#> # ... with 86 more rows

# }