Generate quick scatter correlation plot

Generate scatter correlation plot for 2 variables

scatter_2var_plot(
  data_type = c("variable", "variable.locality"),
  varId,
  year = NULL,
  unitParentId = NULL,
  unitLevel = NULL,
  aggregateId = NULL,
  lang = c("pl", "en"),
  ...
)

Arguments

data_type
A type of data used for generating plot, "variable"(default), "variable.locality"

varId
A vector of 2 variable Id's. Use search_variables or get_variables to find variable id code.

year
A vector of years. If NULL (default) returns data for all available years.

unitParentId
A 12 character NUTS id code of interested unit. (Used only with data_type equal "variable" or "variable.locality")Use search_units or get_units to find unit id code.

unitLevel
A number from 0 to 6, filters the returned unit by its level. (Used only with data_type equal "variable") If NULL (default) no level filters apply. Use get_levels to find more info.

aggregateId
An aggregate id. Use get_aggregates for more info.

lang
A language of returned data, "pl" (default), "en"

...
Other arguments passed on to GET. For example a proxy parameters, see details.

Value

A ggplot2 plot.

Details

Generate quickly `ggplot2` scatter correlation plot, using BDL data.

Scatter plot 2 variables for given units with regression line, confidence interval and correlation coefficient.

To use a proxy to connect, a use_proxy can be passed to GET. For example get_request(id, filters, config = httr::use_proxy(url, port, username, password)).

Examples

# \donttest{
   scatter_2var_plot(data_type = "variable" ,c("415", "60559"), unitLevel = "2")
 # }