Summarize bdl data frame

Prints brief summary with basic statistical functions like mean, standard deviation, variance, min and max for bdl data frame.

# S3 method for bdl
summary(object, ...)

Arguments

object
bdl data frame to summarise

...
other arguments ignored (for compatibility with generic)

Examples

# \donttest{
df <- get_data_by_variable(varId = "3643")
summary(df)

#> # A tibble: 41 x 7
#> # Groups:   id [41]
#>    id           name                           mean    std   min   max variance
#>    <chr>        <chr>                         <dbl>  <dbl> <dbl> <dbl>    <dbl>
#>  1 000000000000 POLSKA                        68.9  11.5      48    87  131.   
#>  2 010000000000 MAKROREGION POŁUDNIOWY         8.35  2.18      4    12    4.77 
#>  3 011200000000 MAŁOPOLSKIE                    3.1   1.62      1     6    2.62 
#>  4 011210000000 REGION MAŁOPOLSKIE             3.1   1.62      1     6    2.62 
#>  5 012400000000 ŚLĄSKIE                        5.25  1.29      2     7    1.67 
#>  6 012410000000 REGION ŚLĄSKIE                 5.25  1.29      2     7    1.67 
#>  7 020000000000 MAKROREGION PÓŁNOCNO-ZACHODNI 10.8   4.23      5    19   17.9  
#>  8 020800000000 LUBUSKIE                       2.1   0.968     1     4    0.937
#>  9 020810000000 REGION LUBUSKIE                2.1   0.968     1     4    0.937
#> 10 023000000000 WIELKOPOLSKIE                  5.1   3.06      2    10    9.36 
#> # ... with 31 more rows

# }